tidy up CU and DIE loading code.
The key thing to realize is that we have two distinct lookup paths:

1. resolving via debug_arange gives us the offset to a CU header
2. resolving via DW_AT_type gives us the offset to a DIE

Remove the diehash in favor of a lazy loading scheme that uses a
binary search tree to locate the CU and its range of DIEs.
This has a dramatic performance gain relative to the old approach of
reading all the DIEs on first access; we go from ~0.4 seconds to trace
wedgie (on a system with libc debug symbols) down to ~0.06 seconds.

This commit introduces a simple slab layer to reduce the amount of time
we're in malloc; there are approx. 7k attributes and 2k dies in each of
libc and libpthread debug symbols.

--HG--
branch : onering