suppress warning about unhandled expr 0x9f (DW_OP_stack_value)
This is a DWARF-4 item that stores the value of the object on
the dwarf expression stack, rather than the address of the object.
We don't have a way to convey this value through to the type resolution
code, which is focused on reading data from memory.

Fixup failed expression evals so that we can render the variable along
with an <optimized out> annotation.

--HG--
branch : onering